Extract some test utilities
